+++ dags/extract/epc_extract.py
@@ -1,4 +1,4 @@
-"""Airflow DAG that schedules the UK Crime extractor in a Kubernetes pod."""
+"""Airflow DAG that schedules the EPC extractor in a Kubernetes pod."""
 
 from datetime import datetime
 
@@ -6,7 +6,7 @@ from airflow.providers.cncf.kubernetes.operators.pod import KubernetesPodOperato
 from airflow.sdk import dag
 from kubernetes.client import models as k8s
 
-from dags.kube_secrets import AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY
+from dags.kube_secrets import AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, EPC_AUTH_TOKEN
 from include.airflow_utils import (
     DATA_IMAGE,
     amber_dags_defaults,
@@ -14,20 +14,18 @@ from include.airflow_utils import (
     clone_and_setup_repo_cmd,
 )
 
-CRIME_EXECUTE_PATH = "extract/crime/current/execute.py"
-
 
 @dag(
-    dag_id="crime_extract",
-    schedule="0 7 1 * *",
+    dag_id="epc_extract",
+    schedule="0 8 1 * *",
     start_date=datetime(2026, 1, 1),
     catchup=False,
     default_args=amber_dags_defaults,
 )
-def crime_extract():
-    crime_extract_cmd = f"""
+def epc_extract():
+    epc_extract_cmd = f"""
     {clone_and_setup_repo_cmd} &&
-    python {CRIME_EXECUTE_PATH}
+    python extract/epc/src/execute.py incremental
     """
 
     KubernetesPodOperator(
@@ -35,12 +33,16 @@ def crime_extract():
         image=DATA_IMAGE,
         image_pull_secrets=[k8s.V1LocalObjectReference(name="amber-ghcr-registry")],
         kubernetes_conn_id="k8s_conn",
-        task_id="crime_extract_task",
-        name="crime-extract-pod",
-        secrets=[AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY],
-        arguments=[crime_extract_cmd],
+        task_id="epc_extract_task",
+        name="epc-extract-pod",
+        secrets=[
+            AWS_ACCESS_KEY_ID,
+            AWS_SECRET_ACCESS_KEY,
+            EPC_AUTH_TOKEN,
+        ],
+        arguments=[epc_extract_cmd],
         do_xcom_push=False,
     )
 
 
-crime_extract()
+epc_extract()